# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Pizza Delivery App
  description: >
    This app supports the Pizza delivery exercise from Perimiter81.
    Basically there are services to add/modify/get and delete users, login services and cart services.
    This enables the processes for supporting what was required on exercise. 
    
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user/isLogin:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: Is a user login? 
      operationId: isLogin
      parameters:
        - in: body
          name: loginUser
          description: The user to check if it is logged in.
          schema:
            $ref: '#/definitions/UsernameRequest' 
      responses:
        "200":
          description: Success, returns true if login and false otherwise
          schema:
            $ref: "#/definitions/TrueFalseResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/cart:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: Add an order to user cart
      operationId: addToCart
      parameters:
        - in: body
          name: cartData
          description: The cart data to add to user
          schema:
            $ref: '#/definitions/CartDataRequest' 
      responses:
        "201":
          description: Added successfully to Cart
          schema:
            $ref: "#/definitions/TrueFalseResponse"
        "400":
          description: Illegal parameters sent
          schema:
            $ref: "#/definitions/CustomErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/cart/checkout:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      description: checkout cart
      operationId: checkout
      parameters:
        - in: body
          name: username
          description: The cart data to add to user
          schema:
            $ref: '#/definitions/UsernameRequest' 
      responses:
        "201":
          description: Added successfully to Cart
          schema:
            $ref: "#/definitions/TrueFalseResponse"
        "400":
          description: Illegal parameters sent
          schema:
            $ref: "#/definitions/CustomErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UsernameRequest:
    required:
      - username
    properties:
      username:
        type: string
...
  TrueFalseResponse:
    required:
      - response
    properties:
      response:
        type: boolean
...
  CustomErrorResponse:
    required:
      - error
    properties:
      error:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  CartDataRequest:
    required:
      - username
      - productName
      - quantity
    properties:
      username:
        type: string
      productName:
        type: string
      quantity:
        type: integer
        default: 1